// src/interfaces/SteamUser.ts
// Generated by https://quicktype.io

export interface SteamUser {
    player: Player;
    items: Items;
}

export interface Items {
    background: Item;
    avatar: Item;
    frame: Item;
}

export interface Item {
    gameId: number | null,
    gameName: string | null,
    src: string | null,
}

export interface Player {
    steamid: string;
    communityvisibilitystate: number;
    profilestate: number;
    personaname: string;
    commentpermission: number;
    profileurl: string;
    avatar: string;
    avatarmedium: string;
    avatarfull: string;
    avatarhash: string;
    lastlogoff: number;
    personastate: number;
    realname: string;
    primaryclanid: string;
    timecreated: number;
    personastateflags: number;
    loccountrycode: string;
}